Updated the player rating metric. now works with ELO rating standards. New sorting method, rating and opponent rating to replace match difficulty. serving highlights now work!

+++ Duo Play/Statistics.swift
@@ -15,5 +15,5 @@ struct Statistics {
     var pointsFor: Int
     var pointsAgainst: Int
     var pointsDifferential: Int
-    var matchDifficulty: Float
+    var matchDifficulty: Int
 }
